shill: don't truncate sequence numbers
Both when handling NetlinkPackets and when managing the pending dump
sequence number shill has previously truncated the 32bit netlink
sequence to a uint16_t.  This causes issues where the untruncated
sequence is used as the key in the map of handlers, meaning that the
handler never gets invoked and the messages are then treated as if they
were broadcast messages.

BUG=chrome-os-partner:46556
TEST=created a program that issued more than 2^16-1 dump requests

Reviewed-on: https://chromium-review.googlesource.com/307110
Commit-Queue: Garret Kelly <gdk@chromium.org>
Tested-by: Garret Kelly <gdk@chromium.org>

+++ shill/net/netlink_packet.h
@@ -38,7 +38,7 @@ class SHILL_EXPORT NetlinkPacket {
   uint16_t GetMessageType() const;
 
   // Get the sequence number from the header.
-  uint16_t GetMessageSequence() const;
+  uint32_t GetMessageSequence() const;
 
   // Returns the remaining (un-consumed) payload length.
   size_t GetRemainingLength() const;
@@ -106,7 +106,7 @@ class SHILL_EXPORT MutableNetlinkPacket : public NetlinkPacket {
   void SetMessageType(uint16_t type);
 
   // Set the sequence number in the header.
-  void SetMessageSequence(uint16_t sequence);
+  void SetMessageSequence(uint32_t sequence);
 
  private:
   DISALLOW_COPY_AND_ASSIGN(MutableNetlinkPacket);
